isl_scheduler.c: isl_sched_graph: allow number of ILP regions to grow
In the current code base, the number of ILP regions required is known
in advance.  When support for intra-statement consecutivity constraints
will be added, the exact number of required regions depends on various
factors that are difficult to analyze in advance.
Even though it would still be possible in this case to compute an upper
bound for the number of required regions, this upper bound may be
fairly large.  Instead, an estimate will be used that may end up
being lower than the actual needs.
Introduce a mechanism to extend the number of ILP regions
at the point where they are needed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>