interface: make choice of class that best matches function name more explicit
If the name of one class is a prefix of the name of another class,
then a given function should be taken to belong to the class
with the longest name that is still a prefix of the function name.
This situation will arise in particular when lists or
subclassing based on a type function will be supported.

Since the keys of the map that stores the classes are sorted
lexicographically, an extension always appears after its prefix and
the original code would pick the right class.
However, it is better to make the choice more explicit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>