export isl_schedule_node as subclasses
This technically breaks compatibility in the bindings because
the already exported isl_schedule_node_band_member_{get,set}_coincident
functions are now considered to belong to the isl::schedule_node_band
class.  However, since there was no mechanism for the user
to check whether the function could be called, it can be reasonably
assumed that nobody is using these functions.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>